pyproject.toml
[tool.poetry]
name = "laboratory"
version = "0.1.0"
description = "Personal hybrid lab"
authors = ["Luke Turner <github@luketurner.org>"]
license = "MIT"

[tool.poetry.scripts]
lab = "laboratory.cli:cli"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
requests = "^2.23.0"
# ansible = "^2.9.12"

pyyaml = "^5.4.1"
toolz = "^0.11.1"
[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
